move environment package to pkg
