+++ pkg/environments/osqueryconf.go
