+++ tls/handlers/post.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/jmpsec/osctrl/environments"
 	"github.com/jmpsec/osctrl/nodes"
+	"github.com/jmpsec/osctrl/pkg/environments"
 	"github.com/jmpsec/osctrl/queries"
 	"github.com/jmpsec/osctrl/settings"
 	"github.com/jmpsec/osctrl/types"
